解决IAR编译工程报警

+++ RT5780Back_RT000_V1.0_20181205/KneadMotor.c
@@ -16,25 +16,25 @@ static void KneadMotor_Set_Pwm_Data(unsigned int ulDuty);
 void KneadMotor_Reset(void)
 {
     //GPIO_PinOutClear(KNEAD_MOTOR_RESET_PORT, KNEAD_MOTOR_RESET_BIT);
-		bKneadReset = 0;
+    bKneadReset = 0;
 }
 
 void KneadMotor_Reset_Cancel(void)
 {
     //GPIO_PinOutSet(KNEAD_MOTOR_RESET_PORT, KNEAD_MOTOR_RESET_BIT);
-		bKneadReset = 1;
+    bKneadReset = 1;
 }
 void KneadMotorBrake_On(void)
 {
     KneadMotor_Set_Pwm_Data(0);
     //GPIO_PinOutClear(KNEAD_MOTOR_DECAY_PORT, KNEAD_MOTOR_DECAY_BIT);
-		bKneadDecay = 0;
+    bKneadDecay = 0;
 }
 void KneadMotorBrake_Off(void)
 {
     KneadMotor_Set_Pwm_Data(0);
     //GPIO_PinOutSet(KNEAD_MOTOR_DECAY_PORT, KNEAD_MOTOR_DECAY_BIT);
-		bKneadDecay = 1;
+    bKneadDecay = 1;
 }
 void KneadMotor_10ms_Int(void)
 {
@@ -43,22 +43,22 @@ void KneadMotor_10ms_Int(void)
     if(nKneadLoss > 2)
     {
         //TIMER_CompareBufSet(KNEAD_MOTOR_TIMER, KNEAD_MOTOR_TIMER_CHANNEL, 0);
-				TIM1_CCR4H = 0;
-				TIM1_CCR4L = 0;
+        TIM1_CCR4H = 0;
+        TIM1_CCR4L = 0;
     }
 }
 //当电机电源电压大于预设值时，设定占空比按比例缩小，否则设定占空比不变
 unsigned int KneadMotor_VoltageAdj(unsigned int setDuty)
 {
     unsigned short adc24;      //此处的电压值已经扩大了100倍
-		unsigned int scale,yushu;
+    unsigned int scale,yushu;
     //ADC_Get_Voltage(ADC_V24,&adc24);   
     adc24 = nVoltage;		
     if(adc24 <= KNEAD_SET_VOLTAGE) 
     {
         return setDuty;        //电压值偏低，返回预设值
     }
-    scale = KNEAD_SET_VOLTAGE *100 / adc24; //计算与设定电压的比例值,百分之scale
+    scale = (unsigned int)KNEAD_SET_VOLTAGE *(unsigned int)100 / adc24; //计算与设定电压的比例值,百分之scale
     setDuty *= scale;
     yushu = setDuty  % 100;
     setDuty /= 100;   //setDuty乘以百分之scale
@@ -71,21 +71,21 @@ void KneadMotor_Set_Pwm_Data(unsigned int ulDuty)
     if(ulDuty == 0)
     {
         //TIMER_CompareBufSet(KNEAD_MOTOR_TIMER, KNEAD_MOTOR_TIMER_CHANNEL, ulDuty);
-				TIM1_CCR4H = 0;
-				TIM1_CCR4L = ulDuty;
+        TIM1_CCR4H = 0;
+        TIM1_CCR4L = ulDuty;
         return;
     }
     if(KneadMotor_Get_Fault() == KNEAD_MOTOR_FAIL) 
     {
         KneadMotor_Reset();
-			  asm("nop");
-			  asm("nop");
+        asm("nop");
+        asm("nop");
         KneadMotor_Reset_Cancel();
     } 
     //ulDuty = KneadMotor_VoltageAdj(ulDuty);
     //TIMER_CompareBufSet(KNEAD_MOTOR_TIMER, KNEAD_MOTOR_TIMER_CHANNEL, ulDuty);
-		TIM1_CCR4H = 0;
-		TIM1_CCR4L = ulDuty;
+    TIM1_CCR4H = 0;
+    TIM1_CCR4L = ulDuty;
     motorStatus = MOTOR_RUN;
 }
 _Bool KneadMotor_IsRun(void)
@@ -93,7 +93,7 @@ _Bool KneadMotor_IsRun(void)
     unsigned int ulDuty;
     //ulDuty = TIMER_CompareBufGet(KNEAD_MOTOR_TIMER, KNEAD_MOTOR_TIMER_CHANNEL);
     ulDuty = TIM1_CCR4L;
-		if(ulDuty > 0) return 1;
+    if(ulDuty > 0) return 1;
     else return 0;
 }
 
@@ -102,18 +102,18 @@ void KneadMotor_ClockRun(void)
     //GPIO_PinOutSet(KNEAD_MOTOR_RESET_PORT, KNEAD_MOTOR_RESET_BIT);
     //GPIO_PinOutSet(KNEAD_MOTOR_PHASE_PORT, KNEAD_MOTOR_PHASE_BIT);
     //GPIO_PinOutClear(KNEAD_MOTOR_DECAY_PORT, KNEAD_MOTOR_DECAY_BIT);
-		bKneadReset = 1;
-		bKneadPhase = 1;
-		bKneadDecay = 0;
+    bKneadReset = 1;
+    bKneadPhase = 1;
+    bKneadDecay = 0;
 }
 void KneadMotor_UnClockRun(void)
 {
     //GPIO_PinOutSet(KNEAD_MOTOR_RESET_PORT, KNEAD_MOTOR_RESET_BIT);
     //GPIO_PinOutClear(KNEAD_MOTOR_PHASE_PORT, KNEAD_MOTOR_PHASE_BIT);
     //GPIO_PinOutClear(KNEAD_MOTOR_DECAY_PORT, KNEAD_MOTOR_DECAY_BIT);
-		bKneadReset = 1;
-		bKneadPhase = 0;
-		bKneadDecay = 0;
+    bKneadReset = 1;
+    bKneadPhase = 0;
+    bKneadDecay = 0;
 }
 void KneadMotor_Break(void)
 {
@@ -121,7 +121,7 @@ void KneadMotor_Break(void)
     if(motorStatus == MOTOR_RUN)
     {
         //GPIO_PinOutSet(KNEAD_MOTOR_DECAY_PORT, KNEAD_MOTOR_DECAY_BIT); //使马达端口处于高阻，此时马达属于惯性滑行
-				bKneadDecay = 1;
+        bKneadDecay = 1;
         motorStatus = MOTOR_STOP_HZ;
         motorBreakTime = 0;
     }
@@ -130,14 +130,14 @@ void KneadMotor_Break(void)
         if(motorBreakTime < MOTOR_STOP_HZ_TIME) return;
     }
     //GPIO_PinOutClear(KNEAD_MOTOR_DECAY_PORT, KNEAD_MOTOR_DECAY_BIT);  //短路马达，保持马达在刹车状态
-		bKneadDecay = 0;
+    bKneadDecay = 0;
     motorStatus = MOTOR_STOP_BREAK;
 }
 
 int KneadMotor_Get_Fault(void)
 {
     if(bKneadFault == 1)
-        return KNEAD_MOTOR_NORMAL;
+    return KNEAD_MOTOR_NORMAL;
     return KNEAD_MOTOR_FAIL;
 }
